Fix profile pic error

+++ controllers/users.controller.js
@@ -38,12 +38,18 @@ exports.uploadImage = [
   async (req, res, next) => {
     try {
       const user = req.user;
-      // Delete file before assign new file in the callback of fs unlink
-      fs.unlink(`public${user.avatar}`, async () => {
+      if (user.avatar !== '/images/avatars/profile-pic.png') {
+        // Delete file before assign new file in the callback of fs unlink
+        fs.unlink(`public${user.avatar}`, async () => {
+          user.avatar = `/images/avatars/${req.file.filename}`;
+          await user.save();
+          res.redirect('/');
+        });
+      } else {
         user.avatar = `/images/avatars/${req.file.filename}`;
         await user.save();
         res.redirect('/');
-      });
+      }
     } catch (err) {
       next(err);
     }
@@ -72,7 +78,6 @@ exports.usersList = async (req, res, next) => {
     const search = req.query.search;
     const users = await usersQueries.searchUsersByUsername(search);
     res.render('includes/search-menu', { users });
-    console.log(users.length);
   } catch (err) {
     next(err);
   }
